+++ templates/default/src/include/src/components/FeedList.vue
@@ -3,7 +3,9 @@ import { ref } from 'vue'
 import FeedItems from './FeedItems.vue'
 import AudioPlayer from './AudioPlayer.vue'
 import FilterBox from './FilterBox.vue'
+import FeedNavigator from './FeedNavigator.vue'
 import EmbedModal from './EmbedModal.vue'
+import IconNotFound from '@/components/icons/IconNotFound.vue'
 import { useFeedsStore } from '@/stores/feeds'
 import { useEmbedStore } from '@/stores/embed'
 import { useAudioStore } from '@/stores/audio'
@@ -19,8 +21,14 @@ const props = defineProps({
     type: Boolean,
     required: true,
   },
+  showNav: {
+    type: Boolean,
+    required: true,
+  },
 })
 
+const showLoadingSpinner = ref(true)
+const totalItemCount = ref(0)
 const expandedFeed = ref(null)
 const expandedArticles = ref([])
 const embedStore = useEmbedStore()
@@ -69,9 +77,24 @@ const generateFirehoseFeed = () => {
     isQuery: false,
   }
 }
+
+const handleLoadingFeed = () => {
+  totalItemCount.value = 0
+  showLoadingSpinner.value = true
+}
+const handleLoadedFeed = (numItems) => {
+  totalItemCount.value += numItems
+  showLoadingSpinner.value = false
+}
 </script>
 
 <template>
+  <div class="loading-spinner" v-if="showLoadingSpinner" />
+  <div id="no-feeds-found-indicator" v-if="totalItemCount === 0 && !showLoadingSpinner">
+    <IconNotFound/>
+    <h2>No feeds found</h2>
+  </div>
+  <FeedNavigator v-if="!props.archived" :show="showNav" />
   <FilterBox />
   <div v-if="filterStore.firehose">
     <FeedItems
@@ -82,9 +105,11 @@ const generateFirehoseFeed = () => {
       :expandedArticles="expandedArticles"
       @expand-article="handleArticleExpand"
       @unexpand-article="handleArticleUnexpand"
+      @feed-loading="handleLoadingFeed"
+      @feed-loaded="handleLoadedFeed"
     />
   </div>
-  <div class="feed-list-wrapper" v-else v-for="feed in feeds" :key="feed.id">
+  <div class="feed-list-wrapper" v-else v-for="(feed, index) in feeds" :key="index">
     <Transition>
       <FeedItems
         :feed="feed"
@@ -93,10 +118,13 @@ const generateFirehoseFeed = () => {
         :expand="showExpandedFeed(feed)"
         :firehose="false"
         :expandedArticles="expandedArticles"
+        :feedIndex="index"
         @expand-feed="handleFeedExpand"
         @unexpand-feed="handleFeedUnexpand"
         @expand-article="handleArticleExpand"
         @unexpand-article="handleArticleUnexpand"
+        @feed-loading="handleLoadingFeed"
+        @feed-loaded="handleLoadedFeed"
       />
     </Transition>
   </div>
@@ -125,5 +153,21 @@ const generateFirehoseFeed = () => {
 .v-leave-to {
   opacity: 0;
   transform: translateY(30px);
+}
+
+#no-feeds-found-indicator {
+  position: absolute;
+  left: 50%;
+  top: 30%;
+  transform: translateX(-50%);
+  transform: translateY(-30%);
+}
+
+#no-feeds-found-indicator svg {
+  width: 60px;
+  height: 60px;
+  display: block;
+  margin: auto;
+  stroke: #c7cfcc;
 }
 </style>