+++ templates/default/src/include/src/components/FilterBox.vue
@@ -23,7 +23,6 @@ const setFirehose = () => {
     firehose: true,
     filterByDays: false,
   })
-
 }
 
 const filters = ref(fStore.filters)
@@ -51,16 +50,16 @@ watch(searchFeedsTerm, (val) => {
 <template>
   <div class="filter-container">
     <span class="filter-box"
-      ><button
-        :class="{ selected: filters.firehose === true }"
-        @click="setFirehose()"
-      >
-      Firehose
+      ><button :class="{ selected: filters.firehose === true }" @click="setFirehose()">
+        Firehose
       </button></span
     >
     <span class="filter-box"
       ><button
-        :class="{ selected: filters.daysBackCount === 1 && filters.filterByDays === true && !filters.firehose }"
+        :class="{
+          selected:
+            filters.daysBackCount === 1 && filters.filterByDays === true && !filters.firehose,
+        }"
         @click="setTimeLimit(1)"
       >
         Last day
@@ -68,7 +67,9 @@ watch(searchFeedsTerm, (val) => {
     >
     <span class="filter-box"
       ><button
-        :class="{ selected: filters.itemCount === 50 && filters.filterByDays === false && !filters.firehose }"
+        :class="{
+          selected: filters.itemCount === 50 && filters.filterByDays === false && !filters.firehose,
+        }"
         @click="setItemLimit(50)"
       >
         Last 50
@@ -76,7 +77,9 @@ watch(searchFeedsTerm, (val) => {
     >
     <span class="filter-box"
       ><button
-        :class="{ selected: filters.itemCount === 20 && filters.filterByDays === false && !filters.firehose }"
+        :class="{
+          selected: filters.itemCount === 20 && filters.filterByDays === false && !filters.firehose,
+        }"
         @click="setItemLimit(20)"
       >
         Last 20
@@ -97,6 +100,7 @@ watch(searchFeedsTerm, (val) => {
   width: 100%;
   padding: 20px 0px 20px 0px;
   margin: 0px 0px 24px 0px;
+  position: relative;
 }
 
 .filter-box {
@@ -118,6 +122,7 @@ watch(searchFeedsTerm, (val) => {
   color: var(--color-text);
   outline: none;
 }
+
 .filter-box button,
 #filter-search span,
 #filter-search button {
@@ -137,6 +142,7 @@ watch(searchFeedsTerm, (val) => {
 #filter-search {
   float: right;
 }
+
 #filter-search input {
   width: 164px;
   height: 29px;
@@ -145,10 +151,16 @@ watch(searchFeedsTerm, (val) => {
   font-weight: normal;
   padding: 0px 0px 0px 6px;
 }
+
 @media (max-width: 640px) {
   .filter-container {
-    margin: 0px;
-    padding: 10px 0px 20px 0px;
+    margin: 0 0 30px 0;
+    position: relative;
+    width: 100%;
+    height: 110px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
   }
   #filter-search {
     float: none;
@@ -158,8 +170,22 @@ watch(searchFeedsTerm, (val) => {
     width: 100%;
     margin-top: 20px;
   }
+
+  #filter-search {
+    float: none;
+    position: absolute;
+    width: 100%;
+    left: 50%;
+    transform: translate(-50%);
+    bottom: -10px;
+  }
+
+  .filter-box {
+    display: inline-block;
+  }
+
   .filter-box button {
-    width: 75px;
+    width: 70px;
   }
 }
 </style>