+++ CHANGELOG.md
@@ -6,13 +6,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.1.1] 2025-01-25
+### Added
+- Add feed navigator
+- Add autoreload option for default template
+### Changed
+- Include query feed articles in RSS feeds
+- Add 'No feeds found' indicator if there are no articles for given view
+- Add loading indicator
+### Fixed
+- Fix audio player not automatically playing when switching between tracks
+
 ## [1.1.0] 2025-01-20
 ### Added
 - Add firehose option to filters
 - Add expansion option for feeds and articles
 - Add total article count to feed items
 - Add custom color to list of available template colors
-- Add comments button to articles from Reddit
+- Add comments button to articles from Reddit and Hackers News
 - Add ability to show article content directly in the app
 - Add isQuery, articleNum, contentLength to list of feed serialized attributes for feed
 - Process all article content removing all superfluous elements and making it more readable